testutil: add delimiter semantics to s3test
Summary:
s3test is a mock framework for S3 for T7892

This diff supports the delimeter/CommonPrefix semantics as below.
Also a unit test for the ListObjectsV2 call is added

From
AWS ListObjectsV2 doc,
"A delimiter is a character you use to group keys.

If you specify a prefix, all keys that contain the same string between the prefix and the first occurrence of the delimiter after the prefix are grouped under a single result element called CommonPrefixes. If you don't specify the prefix parameter, the substring starts at the beginning of the key. The keys that are grouped under the CommonPrefixes result element are not returned elsewhere in the response."

This task is to make the implementation of ListObjectsV2 in testutil.go implement this semantics on the mocked data when ListObjectsV2Input.delimiter is not null.

Reviewers: rck, cbain

Reviewed By: rck

Differential Revision: https://phabricator.grailbio.com/D13975

fbshipit-source-id: 64b6b29