grailfile: add ListDir() command to implementation
Summary:
In order to support large directories and labagent use cases, we
are complementing the implementation.List() command with an
implementation.ListDir() command. This command only accepts
prefixes that correspond to pseudo-directories in S3, ending in /.
It returns a mix of files and pseudo-directories following the
s3 ListObjectsV2 Delimiter semantics.

Also updated localfile.go to support ListDir. Note that stat(d)
does not have size defined if d is a directory.

Following is the comment on

Test Plan:
Added Test.ListDir() to TestAll, and it passes for both local and
s3test.

Reviewers: ysaito, pknudsgaard

Reviewed By: ysaito

Subscribers: schandra, jheart, marius, ayip, cbain

Differential Revision: https://phabricator.grailbio.com/D14809

fbshipit-source-id: 7fa08e6