Simple tool for benchmarking large apps
Summary:
It takes a json config that specifies the commandline, e.g., the bazel target name for the executable, the input files in s3. It then runs the apps in a statically proper fashion.

I added binmetrics and biometrics config.

I wrote this to do a systematic study of go 1.10 and 1.11 performance.

Reviewers: cnicolaou, marius

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D18414

fbshipit-source-id: 7daf66a