gql: handle concurrent updates to s3 cache properly
Summary:
This diff changes the cache to use a shadow-paging like structure.

When multiple scripts try to store&read a cache with the same key concurrently,
one of them would fail because the file's etag would change in the
background. This is somewhat unavoidable because BTSV files aren't guarantee to
be byte-by-byte identical, even if they store the same logical contents. The
reasons include:

  - the columnname-to-columnid mappings may change because they are assigned in FCFS order.
  - the file metadata may change because it embeds the commandline of the GQL invocation.

With this change, when a function wants to produce a cache
s3://grail-query/cache/foo.btsv, it actually writes the contents in
s3://grail-query/cache/foo-randomuid.btsv, where randomuid is a randomly
generated unique ID. After the contents are produced, it creates a "symlink"
file, s3://grail-query/cache/foo.btsv.link that stores the
s3://grail-query/cache/foo-randomuid.btsv as the contents.

As a bonus, this diff also implements cache lookup for some key functions.  I'll
cleanup cache lookup futher in a followup change.

Reviewers: marius, pbohannon

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D18975

fbshipit-source-id: d2c6949