github.com/grailbio/testutil/encryptiontest: allow for greater # of attempts
Summary:
In response to T1997 allow for a greater # of attempts at passing the randomness
test. Note that this test is inherently flaky in the sense that it includes
a statistical test against a p-value, so there is always the chance that it will
fail. Note that it hasn't failed in 18 months, so bumping up the iterations should
work fine. It is intended to catch the case where the low level encryption gets
inadvertently disabled in which case the test will fail deterministically.

Test Plan: unit tests

Reviewers: hfrank, kshashidhar

Reviewed By: kshashidhar

Differential Revision: https://phabricator.grailbio.com/D23120

fbshipit-source-id: 7390db0