s3test/testclient: Make S3 test client behave like AWS S3
Summary:
S3 test client shouldn't do things that AWS S3 wouldn't do.
Specifically, PutObject was adding the content's SHA256
to the headers of the object, which we don't expect S3 to do.

Additionally, the test client should retain and reflect any
metadata set by operations and not treat SHA256 as a special case.

The only additional thing that the test client does that S3 wouldn't
is the following:
If the metadata for an object contains Grail's special key expected
to contain SHA256, then it merely checks that the value matches the
SHA256 of the object's contents.  This done to just protect ourselves
from allowing code that sets something else as the value for that key.

Test Plan:
Not directly tested but through unit tests that use the testclient
for example in D34917 and others.

Reviewers: ysaito, marius, cbain

Reviewed By: ysaito, marius

Subscribers: jcharumilind

Differential Revision: https://phabricator.grailbio.com/D35428

fbshipit-source-id: c63372d