pi-release: adds handling for missing fileset in ref dst
Summary:
The release tool uses filesets to compare reference data,
both in the source directory and the reference release
"destination" directory. The tool compares the contents
of both filesets, to ensure that existing files are not
overwritten (which would break run reproducibility), and
that existing files match files with the same prefix in
the source directory.

The release tool errors out in the following case: copying
reference data to an s3 path which does not contain a fileset.
This occurs frequently when releases with new reference data
versions are made.

This diff allows releasing to paths with missing filesets only
if the path contains no files/directories. If the path does
contain existing files, the release is rejected. To get around
this issue for existing releases, an extra endpoint was added
which can manually add a fileset to a release reference directory.

This diff adds some extra robustness by rejecting the release
if the fileset in the directory does not match the files in
that directory.

Test Plan: Added tests with mock S3 bucket.

Reviewers: pbohannon, spolakh, cbain

Reviewed By: spolakh

Maniphest Tasks: T29534

Differential Revision: https://phabricator.grailbio.com/D39674

fbshipit-source-id: 275a296